refactor(carousel): fetch and filter profiles within component
- Moved the logic for fetching and filtering profiles into the `CarouselCreators` component from the Explore view.
- Updated `types.ts` to remove the now-unused `data` property from `CarouselCreatorsProps`.
- Removed unused imports and redundant profile filtering logic from `explore/view.tsx`.
