+++ src/components/carousel/variants/CarouselCreators.tsx
@@ -1,15 +1,29 @@
 import { memo, FC } from 'react';
-import {PublicationType} from "@lens-protocol/react-web";
+import {ExploreProfilesOrderByType, LimitType, PublicationType, useExploreProfiles} from "@lens-protocol/react-web"
 import { Profile } from '@lens-protocol/api-bindings';
 import CarouselSlide from '@src/components/carousel/components/CarouselSlide';
 import { UserItem } from '@src/components/user-item';
 import CarouselWrapper from './CarouselWrapper';
 import { CarouselCreatorsProps } from '../types';
+import {filterHiddenProfiles} from "@src/utils/profile.ts"
 
  const CarouselCreators: FC<CarouselCreatorsProps> = (params) => {
-  const { data, title, minItemWidth, maxItemWidth } = params;
+  const { title, minItemWidth, maxItemWidth } = params;
 
-  const carouselSettings = {
+   const { data: latestCreatedProfiles } = useExploreProfiles({
+     orderBy: ExploreProfilesOrderByType.LatestCreated,
+     limit: LimitType.Fifty,
+   });
+
+   // FilteredCompletedProfiles is an array of objects, each object has a metadata property and inside exists a displayName en bio property; filter the profiles that not have a displayName and bio property
+   const filtered = latestCreatedProfiles?.filter(
+     (profile: any) => profile.metadata?.displayName && profile.metadata?.bio
+   );
+
+   // Clear ###HIDDEN### profiles
+   const filteredProfiles = filterHiddenProfiles(filtered);
+
+   const carouselSettings = {
     infinite: false,
     slidesToShow: 1,
     speed: 500,
@@ -35,7 +49,7 @@ import { CarouselCreatorsProps } from '../types';
   );
 
   const carouselParams = {
-    data,
+    data: filteredProfiles || [],
     title,
     minItemWidth,
     maxItemWidth,
